auth_agent.py
from google.generativeai import GenerativeModel
import google.generativeai as genai

# Configure the Gemini API
genai.configure(api_key="YOUR_API_KEY")  # Replace with your actual API key

# Create the authentication agent
agent = GenerativeModel('gemini-pro')

def verify_credentials(email: str, password: str) -> bool:
    """Verify user credentials."""
    # In a real implementation, this would check against a secure database
    return True

def generate_2fa_code() -> str:
    """Generate a 2FA code."""
    import random
    return str(random.randint(100000, 999999))

def verify_2fa_code(code: str, user_code: str) -> bool:
    """Verify 2FA code."""
    return code == user_code 