requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.5
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.11.4
typing-extensions>=4.13.2 